FST 95 move taglist
