+++ src/features/integration/IntegrationConfigurationForm.tsx
@@ -10,7 +10,7 @@ import {defaultValues, requiredFields} from "./util/DefaultValues";
 import {toIntegrationConfiguration} from "./util/ToIntegrationConfiguration";
 import AccordionForm from "./components/AccordionForm";
 import {ACCORDION_FORM, IAccordion} from "./types/Accordion";
-import TagList from "./components/dnd/TagList";
+import TagList from "./components/TagList";
 import {HTML5Backend} from "react-dnd-html5-backend";
 import {DndProvider} from "react-dnd";
 import {IIntegrationConfiguration} from "./types/IntegrationConfiguration";
